PWN-3614 - Move serum swap and orca swap parsing to strategy files
