+++ app/src/main/java/org/p2p/wallet/history/strategy/types/UnknownParsingStrategy.kt
@@ -23,5 +23,5 @@ class UnknownParsingStrategy : TransactionParsingStrategy {
         )
     }
 
-    override fun getType(): TransactionDetailsType = TransactionDetailsType.UNKNOWN
+    override fun getType(): TransactionDetailsType = TransactionDetailsType.OTHER
 }
