--- solanaj/src/main/java/org/p2p/solanaj/kits/transaction/parser/OrcaSwapInstructionParser.kt
@@ -1,26 +1,0 @@
-package org.p2p.solanaj.kits.transaction.parser
-
-import org.p2p.solanaj.kits.transaction.SwapDetails
-import org.p2p.solanaj.kits.transaction.network.ConfirmedTransactionRootResponse
-import org.p2p.solanaj.kits.transaction.network.meta.InnerInstructionDetailsResponse
-import org.p2p.solanaj.kits.transaction.network.meta.InstructionResponse
-import java.lang.Exception
-
-object OrcaSwapInstructionParser {
-
-    private class ParsedInstructionDetails(
-        val rootInstruction: InstructionResponse,
-        val innerInstructionDetails: InnerInstructionDetailsResponse?
-    )
-
-
-
-
-
-
-
-
-
-
-
-}
