--- solanaj/src/main/java/org/p2p/solanaj/kits/transaction/parser/SerumSwapInstructionParser.kt
@@ -1,13 +1,0 @@
-package org.p2p.solanaj.kits.transaction.parser
-
-import org.p2p.solanaj.kits.transaction.SwapDetails
-import org.p2p.solanaj.kits.transaction.network.ConfirmedTransactionRootResponse
-import org.p2p.solanaj.kits.transaction.network.meta.InstructionResponse
-import org.p2p.solanaj.programs.SerumSwapProgram
-import java.lang.Exception
-import java.math.BigInteger
-
-object SerumSwapInstructionParser {
-
-
-}
